adds code refactor for tests
